feat: move processing logic out of peer handler to processor
