+++ internal/blocktx/blocktx_mocks.go
@@ -9,8 +9,6 @@ package blocktx
 // from nats_core_client.go
 //go:generate moq -pkg mocks -out ./mocks/mq_client_mock.go . MessageQueueClient
 
-// from peer.go
+// from interface.go
 //go:generate moq -pkg mocks -out ./mocks/peer_mock.go . Peer
-
-// from peer_manager.go
 //go:generate moq -pkg mocks -out ./mocks/peer_manager_mock.go . PeerManager
