+++ internal/blocktx/store/postgresql/register_transactions.go
@@ -16,13 +16,14 @@ func (p *PostgreSQL) RegisterTransactions(ctx context.Context, transactions []*b
 		hashes[i] = transaction.Hash
 	}
 
-	const q = `INSERT INTO blocktx.transactions (hash, is_registered)
-					SELECT hash, TRUE 
-					FROM UNNEST ($1::BYTEA[]) as hash
-				ON CONFLICT (hash) DO UPDATE 
-					SET is_registered = TRUE
-				RETURNING hash, inserted_at
-				`
+	const q = `
+		INSERT INTO blocktx.transactions (hash, is_registered)
+			SELECT hash, TRUE
+			FROM UNNEST ($1::BYTEA[]) as hash
+		ON CONFLICT (hash) DO UPDATE
+			SET is_registered = TRUE
+		RETURNING hash, inserted_at
+	`
 
 	now := p.now()
 	rows, err := p.db.QueryContext(ctx, q, pq.Array(hashes))
