+++ internal/blocktx/store/postgresql/set_block_processing.go
@@ -12,11 +12,10 @@ import (
 )
 
 func (p *PostgreSQL) SetBlockProcessing(ctx context.Context, hash *chainhash.Hash, setProcessedBy string) (string, error) {
-
 	// Try to set a block as being processed by this instance
 	qInsert := `
 		INSERT INTO blocktx.block_processing (block_hash, processed_by)
-		VALUES ($1 ,$2 )
+		VALUES ($1, $2)
 		RETURNING processed_by
 	`
 
@@ -25,6 +24,7 @@ func (p *PostgreSQL) SetBlockProcessing(ctx context.Context, hash *chainhash.Has
 	if err != nil {
 		var pqErr *pq.Error
 
+		// Error 23505 is: "duplicate key violates unique constraint"
 		if errors.As(err, &pqErr) && pqErr.Code == pq.ErrorCode("23505") {
 			err = p.db.QueryRowContext(ctx, `SELECT processed_by FROM blocktx.block_processing WHERE block_hash = $1`, hash[:]).Scan(&processedBy)
 			if err != nil {
@@ -47,7 +47,7 @@ func (p *PostgreSQL) DelBlockProcessing(ctx context.Context, hash *chainhash.Has
 
 	q := `
 		DELETE FROM blocktx.block_processing WHERE block_hash = $1 AND processed_by = $2;
-    `
+  `
 
 	res, err := p.db.ExecContext(ctx, q, hash[:], processedBy)
 	if err != nil {
@@ -64,10 +64,10 @@ func (p *PostgreSQL) DelBlockProcessing(ctx context.Context, hash *chainhash.Has
 func (p *PostgreSQL) GetBlockHashesProcessingInProgress(ctx context.Context, processedBy string) ([]*chainhash.Hash, error) {
 	// Check how many blocks this instance is currently processing
 	q := `
-	SELECT bp.block_hash FROM blocktx.block_processing bp
-	LEFT JOIN blocktx.blocks b ON b.hash = bp.block_hash
-	WHERE b.processed_at IS NULL AND bp.processed_by = $1;
-    `
+		SELECT bp.block_hash FROM blocktx.block_processing bp
+		LEFT JOIN blocktx.blocks b ON b.hash = bp.block_hash
+		WHERE b.processed_at IS NULL AND bp.processed_by = $1;
+	`
 
 	rows, err := p.db.QueryContext(ctx, q, processedBy)
 	if err != nil {
